+++ bootique-metrics-healthchecks/src/main/java/io/bootique/metrics/health/HealthCheckGroup.java
@@ -3,9 +3,9 @@ package io.bootique.metrics.health;
 import java.util.Map;
 
 /**
- * A factory of multiple health checks. A grouping of healthchecks for a given service or module in a single object is
- * often used for the purpose of simplifying injection. At the end these objects are used to populate
- * the central {@link HealthCheckRegistry} singleton.
+ * A container of multiple health checks. A grouping of health checks for a given service or module in a single object
+ * is often used for the purpose of simplifying injection. At the end these objects are used to populate the central
+ * {@link HealthCheckRegistry} singleton.
  *
  * @since 0.8
  */